update: エラー処理、コメント追加

+++ scripts/install-nvim-nightly.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+#
+# Neovim nightlyのインストールスクリプト
+#
+
+# エラー発生時はスクリプトを停止する
+set -e
+
 # インストールディレクトリ
 installDir="$HOME/opt/neovim/nightly"
 # アーカイブ名
@@ -20,12 +27,25 @@ elif [[ $OSTYPE == "linux-gnu"* ]] || [[ $OSTYPE == "freebsd"* ]]; then
   nvimNightlyUrl=https://github.com/neovim/neovim/releases/download/nightly/nvim-linux64.tar.gz
 fi
 
+# URLが空の場合はエラーを出力して終了
+if [[ -z $nvimNightlyUrl ]]; then
+  echo "Error: URL for downloading Neovim nightly build is not set."
+  exit 1
+fi
+
 # 既存のインストールディレクトリを削除
 rm -rf $installDir
 # インストールディレクトリ作成
 mkdir -p $installDir
 # ダウンロード
 wget -O $nvimNightlyArchive $nvimNightlyUrl
+
+# ダウンロードに失敗した場合はエラーを出力して終了
+if [[ $? -ne 0 ]]; then
+  echo "Error: Failed to download Neovim nightly."
+  exit 1
+fi
+
 # アーカイブを展開
 tar xzvf $nvimNightlyArchive -C $installDir --strip-components=1
 # アーカイブは不要なので削除
